ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.16.1
numpy>=1.26.0
Pillow>=10.0.0
pydicom>=2.4.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
